+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This changelog covers what's changed in the MPDX APIs.
 
 ## 4 April 2017
 - TNT Import now imports social media accounts.
+- TNT Import now imports contact envelope_greeting.
 
 ## 3 April 2017
 - Imports API now accepts comma delimited tag_list, instead of tags.